ProgramCreate Page, ProgramPage Created and finished
modified:   app/_layout.jsx
Configured the router to react when "create new program" button is clicked
Configured the router to react when some program is opened

modified:   app/programCreate.jsx
Added option to create a new program the user can choose the name,
the amount of workouts plus the days os the week they will be done,
the Focus of the program(muscle building, powerbuilding... etc)
also added a section to add some notes about the program.

new file:   app/programPage.jsx
Created a new file for the program individual page. Here the user is able to
see the number os days/Workouts, the number of rest days plus the main focus
of the program...All of these informations are read when the program is created or modified
Also add a button to modify the program(days, notes, main focus).
Added a weekly schedule
with the days the user chose organized, these can be acessed by clicking
also information about the workout is display below every day.